src/requirements.txt
flask
requests
Pillow
torch
transformers